Wrap up todays game
I had a lot of fun working on this VR title. VR is an interesting medium
that is incredibly hard to master. I've come a long way over the last
few weeks and I'm incredibly happy with my decision to publish an
Android game every day. The future looks bright!